Fix failure handling for resque-jobs-per-fork when child is KILLed.
When the child job process is killed with the KILL signal, needs to handle
the failure hooks for the child.  Normally resque only processes a single
job in the forked child, so it uses the job instance that is already in
memory in the parent process. However, this with resque-jobs-per-fork, this
is just the first job processed by the child, since it pops more jobs off
the queue.  Instead we need to get the job payload from redis.